perf(frontend): speed up extractNote retry and AddRelationship finalize
Merge preview+retry into one mount; merge navigate on/off relationship
finalize paths and extract finalize test support.
